+++ share/qt/clean_mac_info_plist.py
@@ -9,10 +9,10 @@ from datetime import date
 bitcoinDir = "./";
 
 inFile     = bitcoinDir+"/share/qt/Info.plist"
-outFile    = "Catcoin-Qt.app/Contents/Info.plist"
+outFile    = "FuMPcoin-Qt.app/Contents/Info.plist"
 version    = "unknown";
 
-fileForGrabbingVersion = bitcoinDir+"catcoin-qt.pro"
+fileForGrabbingVersion = bitcoinDir+"fumpcoin-qt.pro"
 for line in open(fileForGrabbingVersion):
 	lineArr = line.replace(" ", "").split("=");
 	if lineArr[0].startswith("VERSION"):