+++ doc/release-notes.md
@@ -6,7 +6,7 @@
 0.8.6.2 changes
 =============
 
-- Catcoin fork and rebranding
+- FuMPcoin fork and rebranding
 
 0.8.6.1 changes
 =============